add support for fixed crop for zooniverse grid subjects, add tests
