+++ galaxy_datasets/transforms.py
@@ -1,15 +1,12 @@
 import logging
 
 from omegaconf import DictConfig
-import torch
-import numpy as np
-import albumentations as A
-from albumentations.pytorch import ToTensorV2
-import torchvision.transforms.v2 as T
-import PIL
 
+import torch
+import torchvision.transforms.v2 as T
 from torchvision.transforms import InterpolationMode
 
+from galaxy_datasets.transforms_custom_torchvision import FixedCrop
 
 class GalaxyViewTransform():
     # most powerful and most general transform options
@@ -33,6 +30,9 @@ class GalaxyViewTransform():
         antialias = True
         
         transform = []
+
+        if cfg.pil_to_tensor:  # galaxydataset loads as PIL image, tensor is faster to work with
+            transform.append(T.PILToTensor())
         
         if cfg.greyscale:
             transform.append(T.Grayscale(num_output_channels=1))
@@ -41,6 +41,13 @@ class GalaxyViewTransform():
         # disabled for beluga
         # 'torchvision.transforms.v2' has no attribute 'ToImage'
         # requires 0.16.0, beluga has 0.15.2
+
+        if cfg.fixed_crop:
+            # fixed crop with specified corners, useful to pick a single image from Zooniverse grid subjects
+            transform.append(FixedCrop(
+                **cfg.fixed_crop  # lower_left_x, lower_left_y, upper_right_x, upper_right_y
+            ))
+            logging.info(f'Using FixedCrop, {cfg.fixed_crop}')
 
         # shear/perspective transforms happen before any cropping
         if cfg.random_affine:  # will use this for rotation and off-center zoom/crop
@@ -116,7 +123,7 @@ class GalaxyViewTransform():
             )
 
         # finally, shift to 0-1 float32 before hitting model etc
-        # transform.append(T.ToDtype(torch.float32, scale=True))
+        transform.append(T.ToDtype(torch.float32, scale=True))
         # transform.append(T.ConvertImageDtype(torch.float32))
 
         self.transform = T.Compose(transform)
@@ -130,6 +137,8 @@ class GalaxyViewTransform():
 def default_view_config():
     # for debugging
      return DictConfig(dict(
+        pil_to_tensor=True,
+        fixed_crop=False,
         output_size=224,
         greyscale=False,
         interpolation_method='bilinear',
@@ -155,6 +164,8 @@ def default_view_config():
 
 def minimal_view_config():
     return DictConfig(dict(
+        pil_to_tensor=True,
+        fixed_crop=False,
         output_size=224,
         greyscale=False,
         interpolation_method='bilinear',
@@ -173,7 +184,10 @@ def minimal_view_config():
 
 def fast_view_config():
     return DictConfig(dict(
+        pil_to_tensor=True,
+        fixed_crop=False,
         output_size=224,
+        greyscale=False,
         interpolation_method='bilinear',
         random_affine=False,
         random_perspective=False,
@@ -187,245 +201,5 @@ def fast_view_config():
         elastic_prob=0.
     ))
 
-
-"""And below, the older transform options"""
-
-
-def default_transforms(
-    crop_scale_bounds=(0.7, 0.8),
-    crop_ratio_bounds=(0.9, 1.1),
-    initial_center_crop=None,
-    resize_after_crop=224, 
-    pytorch_greyscale=False,
-    to_float=True,  # set to True when loading images directly, False via webdatasets (which normalizes to 0-1 on decode)
-    to_tensor=True
-    ) -> A.Compose:
-
-    transforms_to_apply = base_transforms(pytorch_greyscale)
-
-    if initial_center_crop:
-        transforms_to_apply += [
-            A.CenterCrop(
-                height=initial_center_crop,  # initial crop
-                width=initial_center_crop,
-                # always_apply=True
-            )
-        ]
-
-    transforms_to_apply += [
-        # A.ToFloat(),
-        # anything outside of the original image is set to 0.
-        A.Rotate(limit=180, interpolation=1,
-                    always_apply=True, border_mode=0, value=0),
-        A.RandomResizedCrop(
-            size=(resize_after_crop, resize_after_crop), # after crop resize
-            scale=crop_scale_bounds,  # crop factor
-            ratio=crop_ratio_bounds,  # crop aspect ratio
-            interpolation=1,  # This is "INTER_LINEAR" == BILINEAR interpolation. See: https://docs.opencv.org/3.4/da/d54/group__imgproc__transform.html
-            # always_apply=True
-        ),  # new aspect ratio
-        A.VerticalFlip(p=0.5)
-    ]
-    if to_float:
-        transforms_to_apply += [A.ToFloat(max_value=255.0, always_apply=True)]
-
-    if to_tensor:
-        transforms_to_apply += [ToTensorV2(always_apply=True)]
-
-    return A.Compose(transforms_to_apply)
-
-
-
-def minimal_transforms(
-    resize_after_crop=224, 
-    pytorch_greyscale=False
-    ) -> A.Compose:
-    # for testing how much augmentations slow training / picking CPU to allocate
-    transforms_to_apply = base_transforms(pytorch_greyscale)
-    transforms_to_apply += [
-        A.CenterCrop(
-            height=resize_after_crop,
-            width=resize_after_crop,
-            always_apply=True
-        )
-    ]
-    return A.Compose(transforms_to_apply)
-
-
-def fast_transforms(
-    pytorch_greyscale=False,
-    resize_after_crop=224
-    ):
-    # middle ground between default and minimal transforms
-    # faster than default because we avoid interpolation
-    # better than minimal because we have some rotation/flip, and use random (non-central) crop
-    # should only be used for proper training if you already resize the images
-    # such that the resize_after_crop FoV makes sense (as this is just cropping)
-    # for 0.75x=224, x=300, so save at 300x300 pixels!
-    assert resize_after_crop == 224  # check user isn't attempting to change this
-    transforms_to_apply = base_transforms(pytorch_greyscale)
-    transforms_to_apply += [
-    #     A.RandomCrop(
-    #         height=resize_after_crop,
-    #         width=resize_after_crop,
-    #         always_apply=True
-    #     ),
-        A.Flip(),
-        A.RandomRotate90()
-    ]
-    return A.Compose(transforms_to_apply)
-
-
-def base_transforms(pytorch_greyscale):
-    if pytorch_greyscale:
-        return [
-            A.Lambda(
-                name="ToGray", image=ToGray(reduce_channels=True) #, always_apply=True
-            )
-        ]
-    else:
-       return [
-            A.Lambda(name="RemoveAlpha", image=RemoveAlpha()) #, always_apply=True)
-        ]
-        
-
-
-def astroaugmentation_transforms(
-    resize_size: int,
-    shift_limit: float,
-    scale_limit: float,
-    rotate_limit: float,
-    p_channelwise_dropout: float,
-    p_elastic=0,
-    elastic_alpha=1,
-    elastic_sigma=1,
-    channelwise_dropout_max_fraction=0.2,
-    channelwise_dropout_min_length=10,
-    channelwise_dropout_max_holes=100,
-    pytorch_greyscale=False,
-) -> A.Compose:
-    """
-    Astrophysically-inspired image transforms from AstroAugmentations (Bowles in prep.)
-    Intended to mimic common distortions to optical telescope images.
-    
-    Sequentially:
-    - ElasticTransform (optional) from https://ieeexplore.ieee.org/document/1227801
-    - ChannelWiseDropout (good for Decals) from AstroAugmentations
-    - ShiftScaleRotate instead of RandomResizedCrop for more control
-    - Flip along x or y axis
-
-    Args:
-        resize_size (int): resolution
-        shift_limit (float): max relative shift factor. 0.3 can have the galaxy pretty much
-                            at the edge *without scaling*
-        scale_limit (float):  max relative max zoom *in* factor
-        rotate_limit (float): max rotate angle in degrees
-        p_channelwise_dropout (float): _description_
-        p_elastic (int, optional): how much to scale the random field. Defaults to 0.
-        elastic_alpha (int, optional): std. of the Gauss. kernel blurring the random field. Defaults to 1.
-        elastic_sigma (int, optional):  max area fraction to be affected. Defaults to 1.
-        channelwise_dropout_max_fraction (float, optional): _description_. Defaults to 0.2.
-        channelwise_dropout_min_length (int, optional):  minimum length of the hole in pixels. Defaults to 10. May be horizontal or vertical (width or height)
-        channelwise_dropout_max_holes (int, optional): _description_. Defaults to 100.
-        pytorch_greyscale (bool, optional): _description_. Defaults to False.
-
-    Returns:
-        A.Compose: _description_
-    """
-    # wrapped in Try/Except to avoid making AstroAugmentations a package requirement for only this optional transform
-    try:
-        from AstroAugmentations import image_domain  # type: ignore
-    except ImportError:
-        raise ImportError(
-            'Trying to use astroaugmentation_transforms but AstroAugmentations is not installed\n \
-            Please install via git. See instructions at https://github.com/mb010/AstroAugmentations#quick-start'
-        )
-    from albumentations.pytorch import ToTensorV2  # also required pytorch
-
-    transforms_to_apply = [
-        A.Lambda(name="RemoveAlpha", image=RemoveAlpha(), always_apply=True)
-    ]
-
-    if pytorch_greyscale:
-        transforms_to_apply += [
-            A.Lambda(
-                name="ToGray", image=ToGray(reduce_channels=True), always_apply=True
-            )
-        ]
-
-    transforms_to_apply += [
-        A.LongestMaxSize(max_size=resize_size),
-        A.ElasticTransform(
-            alpha=elastic_alpha,
-            sigma=elastic_sigma,
-            alpha_affine=0,
-            interpolation=1,
-            border_mode=1,
-            value=0,
-            p=p_elastic,
-        ),
-        A.Lambda(
-            name="MissingData",
-            image=image_domain.optical.ChannelWiseDropout(
-                max_fraction=channelwise_dropout_max_fraction,
-                # max is up to channelwise_dropout_max_fraction
-                min_width=channelwise_dropout_min_length,
-                min_height=channelwise_dropout_min_length,
-                max_holes=channelwise_dropout_max_holes,
-                channelwise_application=True,
-            ),
-            p=p_channelwise_dropout,
-        ),
-        A.ShiftScaleRotate(
-            shift_limit=shift_limit,
-            scale_limit=(0, scale_limit), # type: ignore
-            rotate_limit=rotate_limit, # type: ignore
-            interpolation=2,
-            border_mode=0,
-            p=1,
-        ),
-        # TODO maybe add sersic/gaussian
-        A.Flip(p=0.5),
-        ToTensorV2(),
-    ]
-
-    return A.Compose(transforms_to_apply)
-
-
-# albumentations version of GrayscaleUnweighted
-class ToGray():
-
-    def __init__(self, reduce_channels=False):
-        self.reduce_channels = reduce_channels
-
-    def forward(self, img):
-        if len(img.shape) == 2:  # saved to disk as greyscale already, with no channel
-            img = np.expand_dims(img, axis=2) # add channel=1 dimension
-        # print(img.shape)
-        if self.reduce_channels:
-            return img.mean(axis=2, keepdims=True)
-        else:
-            return img.mean(axis=2, keepdims=True).repeat(3, axis=2)
-
-    def __call__(self, image, **kwargs):
-        return self.forward(image)
-
-class RemoveAlpha():
-
-    def __init__(self):
-        # some png images have fourth alpha channel with value of 255 everywhere (i.e. opaque). averaging over this adds incorrect offset
-        pass
-
-    def forward(self, img):
-        if img.shape[2] == 4:
-            return img[:, :, :3]
-        return img
-
-    def __call__(self, image, **kwargs):
-        return self.forward(image)
-
-
-
-# def interpolation_lookup(interpolation_str='bilinear'):
-#         # PIL.Image.bilinear
-#         return getattr(PIL.Image, interpolation_str.upper())
+# for now, will deprecate
+from galaxy_datasets.transforms_albumentations import minimal_transforms, fast_transforms, default_transforms, base_transforms
